Enable aggressive static analysis flags - resolve resulting issues
Enable -Wall -Wextra and -Werror (as well as a few more useful lints)
and then fix everything to get it compiling.

* The snprintf statements in mm_carrier.c and mm_carrier_mtrl.c seem
  to leave a space for the null terminator, but snprintf guarantees a
  null terminator, so you will get truncation!
* A number of places had for loops with `int i` but were compared
  against the output of `sizeof` - switched them to `size_t i`
* There were a number of places which used `%llx`, `%lld`, and `%llu`
  to print out uint64_t or int64_t types - switch these to their
  standard `PRIx64`, `PRId64`, and `PRIu64` counterparts
* Add in a few missing function prototypes
* Remove const from some functions which claimed their return type was
  const, which didn't make sense since the functions returned basic
  types by copy

+++ mm_areacode.c
@@ -68,7 +68,7 @@ int main(int argc, char *argv[]) {
 
     fclose(instream);
 
-    for (int i = 0; i < sizeof(dlog_mt_npa_sbr_table_t) - 1; i++) {
+    for (size_t i = 0; i < sizeof(dlog_mt_npa_sbr_table_t) - 1; i++) {
         uint8_t c;
         uint8_t flags0, flags1;
